Convert hw contraints for disk/network to list
Actually disk and network now are lists in the spec:

https://tmt.readthedocs.io/en/stable/spec/plans.html#hardware

so make sure we make this detail to user invisible, let him
just specify `--hardware disk.size` and disk will be a list
by default. Not a silver bullet, but makes the currently
expected experience on easy cmdline specification actually
working.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>