request: do not set the timeout if not force
The default timeout is different per ranch, see docs/root!214.
The original code was using the hardcoded timeout 12 hours to check
if it should force the timeout when doing the request. That meant
if on RH ranch, where the timeout is 24h by default, and the user
set the timeout to 12h, it did not force that, and the 24 default
was used :)

This is very confusing, so change the forcing based on the parameter
source, if it is via commmand line, always force it.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>