request/restart: fix --test and --test-filter usage with --reserve
When these options are used with `--reserve` we need to inject
the reservation task to the mix.

For `--test one|two|three`, we add `|/testing-farm/reserve-system`
to the test regex.

For `--test-filter 'tag:smoke & name:/sanity', we add OR filter
`tag:smoke & name:/sanity | name:/testing-farm/reserve-system`.

For test filter documentation see:
https://fmf.readthedocs.io/en/latest/modules.html#fmf.filter

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>